Added Consoloid.FileList.Dialog.Welcome, which might eventually show some device info, right now it starts a file list view
Also it's a workaround for how the first dialog does not have an expression, therefore it can not set and expression reference.
Also changed the default folder of the file list view to "/", because Android does not have a "/tmp"... actual Android implementaion might set it to a more public folder (eg. /sdcard or something)